requirements.txt
flask==3.0.2
gradio==4.44.1
requests==2.32.3
huggingface_hub<0.24
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
defusedxml==0.7.1
gunicorn==22.0.0
waitress==3.0.0
a2wsgi==1.10.10
SQLAlchemy==2.0.32
alembic==1.13.2
playwright==1.49.0

pytest==8.2.0
pytest-cov==5.0.0
ruff==0.5.6
black==24.8.0
